refactor: move Run() to public API
