+++ google/cloud/functions/integration_tests/cloud_event_integration_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/functions/internal/framework.h"
+#include "google/cloud/functions/framework.h"
 #include "google/cloud/functions/http_response.h"
 #include <boost/asio/connect.hpp>
 #include <boost/asio/ip/tcp.hpp>
